+++ sdk/azidentity/logging.go
@@ -14,14 +14,14 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/internal/log"
 )
 
-// LogCredential entries contain information about authentication.
+// EventCredential entries contain information about authentication.
 // This includes information like the names of environment variables
 // used when obtaining credentials and the type of credential used.
-const LogCredential log.Classification = "Credential"
+const EventCredential log.Event = "Credential"
 
 // log environment variables that can be used for credential types
 func logEnvVars() {
-	if !log.Should(LogCredential) {
+	if !log.Should(EventCredential) {
 		return
 	}
 	// Log available environment variables
@@ -42,25 +42,25 @@ func logEnvVars() {
 		envVars = append(envVars, "AZURE_CLI_PATH")
 	}
 	if len(envVars) > 0 {
-		log.Writef(LogCredential, "Azure Identity => Found the following environment variables:\n\t%s", strings.Join(envVars, ", "))
+		log.Writef(EventCredential, "Azure Identity => Found the following environment variables:\n\t%s", strings.Join(envVars, ", "))
 	}
 }
 
 func logGetTokenSuccess(cred azcore.TokenCredential, opts policy.TokenRequestOptions) {
-	if !log.Should(LogCredential) {
+	if !log.Should(EventCredential) {
 		return
 	}
 	msg := fmt.Sprintf("Azure Identity => GetToken() result for %T: SUCCESS\n", cred)
 	msg += fmt.Sprintf("\tCredential Scopes: [%s]", strings.Join(opts.Scopes, ", "))
-	log.Write(LogCredential, msg)
+	log.Write(EventCredential, msg)
 }
 
 func logCredentialError(credName string, err error) {
-	log.Writef(LogCredential, "Azure Identity => ERROR in %s: %s", credName, err.Error())
+	log.Writef(EventCredential, "Azure Identity => ERROR in %s: %s", credName, err.Error())
 }
 
 func logMSIEnv(msi msiType) {
-	if !log.Should(LogCredential) {
+	if !log.Should(EventCredential) {
 		return
 	}
 	var msg string
@@ -74,11 +74,11 @@ func logMSIEnv(msi msiType) {
 	default:
 		msg = "Azure Identity => Managed Identity environment: Unknown"
 	}
-	log.Write(LogCredential, msg)
+	log.Write(EventCredential, msg)
 }
 
 func addGetTokenFailureLogs(credName string, err error, includeStack bool) {
-	if !log.Should(LogCredential) {
+	if !log.Should(EventCredential) {
 		return
 	}
 	stack := ""
@@ -86,5 +86,5 @@ func addGetTokenFailureLogs(credName string, err error, includeStack bool) {
 		// skip the stack trace frames and ourself
 		stack = "\n" + diag.StackTrace(3, 32)
 	}
-	log.Writef(LogCredential, "Azure Identity => ERROR in GetToken() call for %s: %s%s", credName, err.Error(), stack)
+	log.Writef(EventCredential, "Azure Identity => ERROR in GetToken() call for %s: %s%s", credName, err.Error(), stack)
 }